Fix grid implementation! Had a big bug.
I assumed that the Grid always was a square but it wasn't. The x and y
lengths were frequently different.

I changed the internal representation of the grid to use nested lists,
fixed the tests, found a bug in one of my tests, and now
smoke_low_basins is also passing.